pyproject.toml
[tool.poetry]
name = "fabric"
version = "1.2.0"
description = "Fabric - AI framework for human augmentation"
authors = [
    "Daniel Miessler <https://github.com/danielmiessler>",
    "Jonathan Dunn <https://github.com/xssdoctor>",
    "Scott Behrens <https://github.com/sbehrens>",
]
packages = [
    { include = "installer"}
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = "^0.30.11"
unstructured = "0.10.25"
pyowm = "3.3.0"
tools = "^0.1.9"
langchain-community = "^0.0.25"
google-api-python-client = "^2.120.0"
isodate = "^0.6.1"
youtube-transcript-api = "^0.6.2"
pydub = "^0.25.1"
ollama = "^0.1.7"
anthropic = "^0.18.1"
pyperclip = "^1.8.2"
python-dotenv = "1.0.0"
jwt = "^1.3.1"
flask = "^3.0.2"
helpers = "^0.2.0"
praisonai = "^0.0.18"
google-generativeai = "^0.5.3"

[tool.poetry.group.cli.dependencies]
pyyaml = "^6.0.1"
requests = "^2.32.0"
python-socketio = "^5.11.0"
websocket-client = "^1.7.0"
flask = "^3.0.2"
flask-sqlalchemy = "^3.1.1"
flask-login = "^0.6.3"
flask-jwt-extended = "^4.6.0"
python-dotenv = "1.0.0"
openai = "^1.11.0"
flask-socketio = "^5.3.6"
flask-sock = "^0.7.0"
gunicorn = "^22.0.0"
gevent = "^23.9.1"
httpx = ">=0.25.2,<0.26.0"
tqdm = "^4.66.3"

[tool.poetry.group.server.dependencies]
requests = "^2.32.0"
openai = "^1.12.0"
flask = "^3.0.2"
python-dotenv = "1.0.0"
jwt = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# Creates python "binaries"
[tool.poetry.scripts]
fabric = 'installer:cli'
fabric-api = 'installer:run_api_server'
fabric-webui = 'installer:run_webui_server'
ts = 'installer:main_ts'
yt = 'installer:main_yt'
save = 'installer:main_save'
